refactor: use date utils from common instead
