+++ src/apps/members/apps/add/app.ts
@@ -8,8 +8,7 @@ import express from "express";
 import { getRepository } from "typeorm";
 
 import { hasSchema, isSuperAdmin } from "@core/middleware";
-import { wrapAsync } from "@core/utils";
-import { createDateTime } from "@core/utils/date";
+import { createDateTime, wrapAsync } from "@core/utils";
 
 import MembersService from "@core/services/MembersService";
 import OptionsService from "@core/services/OptionsService";
