+++ src/core/utils/index.ts
@@ -128,6 +128,18 @@ export function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
+export function createDateTime(date: string, time: string): Date;
+export function createDateTime(
+  date: string | undefined,
+  time: string | undefined
+): Date | null;
+export function createDateTime(
+  date: string | undefined,
+  time: string | undefined
+): Date | null {
+  return date && time ? new Date(date + "T" + time) : null;
+}
+
 interface PgError {
   code: string;
   detail: string;
